backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
yfinance==0.2.50
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
sqlmodel==0.0.21
bcrypt==4.2.1
resend==2.10.0
beautifulsoup4==4.12.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
